1.Added polonix public client 2.Removed fiatconverter 3.All USD-BTC order books are up and running
