+++ arbitrage/public_markets/GDAX/__init__.py
@@ -1,0 +1,1 @@
+# Placeholder, no content
