+++ arbitrage/public_markets/_poloniex.py
@@ -3,26 +3,30 @@ Created on Oct 4, 2017
 
 @author: stan4
 '''
-import requests
 import json
+import urllib.request
+import urllib.error
+import urllib.parse
+import logging
 from .market import Market
 
 class Poloniex(Market):
-    def __init__(self, currency, product_id):
+    api_url = 'https://poloniex.com/public'
+    
+    def __init__(self, currency, currency_pair):
         super().__init__(currency)
-        self.product_id = product_id
+        self.currency_pair = currency_pair
         self.update_rate = 30
-        self.client = PublicClient()
 
     def update_depth(self):
-#         url = 'https://api.exchange.coinbase.com/products/%s/book?level=2' % self.code
-#         req = urllib.request.Request(url, headers={
-#             "Content-Type": "application/x-www-form-urlencoded",
-#             "Accept": "*/*",
-#             "User-Agent": "curl/7.24.0 (x86_64-apple-darwin12.0)"})
-#         res = urllib.request.urlopen(req)
-#         depth = json.loads(res.read().decode('utf8'))     
-        depth = self.client.get_product_order_book(self.product_id, 2)
+        url = self.api_url + '?command=returnOrderBook&currencyPair={}'.format(self.currency_pair)
+        req = urllib.request.Request(url, headers={
+            "Content-Type": "application/x-www-form-urlencoded",
+            "Accept": "*/*",
+            "User-Agent": "curl/7.24.0 (x86_64-apple-darwin12.0)"})
+        res = urllib.request.urlopen(req)
+        depth = json.loads(res.read().decode('utf8'))
+#         logging.warning(json.dumps(depth))
         self.depth = self.format_depth(depth)
 
     def sort_and_format(self, l, reverse=False):
