--- arbitrage/public_markets/okcoincny.py
@@ -1,5 +1,0 @@
-from ._okcoin import OKCoin
-
-class OKCoinCNY(OKCoin):
-    def __init__(self):
-        super().__init__("CNY", "btc_cny")
